Introduced thread/block policies
* each arch and sub-arch can define their own block/grid
distributions by defining num. operations per block

Introduced VM

* vm unites hw. description, floating point type for code generation
and hw specific lexic. This makes it easy to pass all three between
objects

a Bug fix

* removed an imported module from chainforge

+++ benchmarks/bandwidth/kernel.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-void copyData(float *To, float *From, size_t size, long blocks, long threads, void* stream);
+void copyData(float *To, float *From, size_t size, size_t blocks, size_t threads, void* stream);
 
 #endif //SIMPLE_BANDWIDTH_TEST_KERNEL_H
